+++ CHANGES.md
@@ -1,6 +1,8 @@
 1.6.0
 
-- added support for parents specified as a std::vector of tasks
+- added support for parents provided as a std::vector of tasks
+- ensured that abandoned tasks are canceled
+- fixed bug where tasks were still canceled after reset
 
 1.5.0